chore(support): satisfy CI — ruff format + split oversized test file
- ruff-format service.py and test_support_api.py (format-check was failing).
- Move the upload-target idempotency integration test into
  test_support_upload_idempotency.py so test_support_api.py stays under the
  600-line repo-shape limit (it was 707; back to 590) without allowlisting.
